# docs/openapi.yaml
openapi: 3.0.3
info:
    title: Goveto Edge Control API
    version: 0.1.0
    description: Control-plane API for managing edge clusters, nodes, sites, certificates, publish, purge and analytics.
servers:
    - url: /
...
paths:
...
    /api/v1/clusters/{cluster_id}/analytics/summary:
        get:
            operationId: summary
            summary: Analytics summary
            description: Aggregate request/traffic summary for a time range (from/to).
            tags:
                - analytics
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: from
                  in: query
                  schema:
                    type: string
                    format: date-time
                    example: "2026-01-02T15:04:05Z"
                - name: site_id
                  in: query
                  schema:
                    type: string
                - name: to
                  in: query
                  schema:
                    type: string
                    format: date-time
                    example: "2026-01-02T15:04:05Z"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
...
components:
...
    securitySchemes:
        cookie_session:
            type: apiKey
            in: cookie
            name: session